cassis: split monolithic __init__ into multiple files
