+++ antismash/detection/cassis/__init__.py
@@ -3,25 +3,21 @@
 
 """Implementation of the CASSIS method for the motif-based prediction of SM gene clusters"""
 
-from collections import defaultdict
-import csv
 import logging
 import os
 import shutil
-from typing import Any, Dict, List, Optional, Set, Tuple
-from xml.etree import cElementTree as ElementTree
+from typing import Any, Dict, List
 
-from Bio import SeqIO
-from Bio.SeqRecord import SeqRecord
 from Bio.SeqFeature import SeqFeature
 
 from antismash.common import path, module_results
-from antismash.common.secmet.feature import ClusterBorder, Feature, FeatureLocation, GeneFunction, Gene
+from antismash.common.secmet.feature import ClusterBorder, Feature, FeatureLocation, GeneFunction
 from antismash.common.secmet.record import Record
 from antismash.config.args import ModuleArgs
 
-from .runners import run_fimo, run_meme
-from .promoters import Promoter, CombinedPromoter, get_promoters, DuplicatePromoterError, \
+from .cluster_prediction import get_predictions_for_anchor, ClusterPrediction
+from .pairings import PROMOTER_RANGE
+from .promoters import CombinedPromoter, get_promoters, DuplicatePromoterError, \
                        write_promoters_to_file
 
 NAME = "cassis"
@@ -115,30 +111,6 @@ class InvalidLocationError(Exception):
     pass
 
 
-class Pairing:
-    """ A common component throughout CASSIS is a range given by number of
-        promoters upstream and number of promoters downstream. This base class
-        is used to reduce duplication and simplify string generation for labels.
-    """
-    def __init__(self, plus: int, minus: int) -> None:
-        self.plus = int(plus)
-        self.minus = int(minus)
-
-    @property
-    def pairing_string(self) -> str:
-        """ A string representing the range of this pairing. """
-        return "+{:02d}_-{:02d}".format(self.plus, self.minus)
-
-    def __str__(self) -> str:
-        return self.pairing_string
-
-
-# all possible promoter sets for motif detection
-# plus --> include <plus> promoters downstream the anchor gene's promoter
-# minus --> include <minus> promoters upstream the anchor gene's promoter
-PROMOTER_RANGE = [Pairing(plus, minus) for plus in range(16) for minus in range(16) if plus + minus >= 3]
-
-
 def check_prereqs():
     """Check for prerequisites"""
     failure_messages = []
@@ -148,63 +120,6 @@ def check_prereqs():
         # TODO: Check binary version here
 
     return failure_messages
-
-
-class Motif(Pairing):
-    """ An individual motif, tracks a location as a plus/minus pair, an evalue (score)
-        and any hits.
-    """
-    def __init__(self, plus: int, minus: int, score: Optional[float] = None,
-                 hits: Optional[Dict[str, int]] = None) -> None:
-        super().__init__(plus, minus)
-        self._score = None  # type: Optional[float]
-        if score:
-            self.score = score
-        self.seqs = []  # type: List[str]
-        self.hits = defaultdict(lambda: 0)  # type: Dict[str, int]
-        if hits:
-            for key, val in hits.items():
-                self.hits[str(key)] = int(val)
-
-    @property
-    def score(self) -> Optional[float]:
-        """ The score/evalue of a motif """
-        return self._score
-
-    @score.setter
-    def score(self, score: float):
-        self._score = float(score)
-
-    def __eq__(self, other: Any) -> bool:
-        return (isinstance(other, Motif)
-                and self.plus == other.plus
-                and self.minus == other.minus
-                and self._score == other.score
-                and self.seqs == other.seqs
-                and self.hits == other.hits)
-
-    def __repr__(self) -> str:
-        return "Motif(%s, score=%s, hits=%s)" % (self.pairing_string, self.score, self.hits)
-
-
-class Island:
-    """ A container for an island between two promoters with a motif. """
-    def __init__(self, start: Promoter, end: Promoter, motif: Motif) -> None:
-        assert isinstance(start, Promoter)
-        self.start = start
-        assert isinstance(end, Promoter)
-        self.end = end
-        assert isinstance(motif, Motif)
-        self.motif = motif
-
-    def __eq__(self, other: Any) -> bool:
-        return (isinstance(other, Island)
-                and self.start == other.start
-                and self.end == other.end
-                and self.motif == other.motif)
-
-    def __repr__(self) -> str:
-        return "Island(start=%s, end=%s, motif=%s)" % (self.start, self.end, self.motif)
 
 
 def detect(record: Record, options) -> List[ClusterBorder]:
@@ -251,57 +166,12 @@ def detect(record: Record, options) -> List[ClusterBorder]:
     cluster_predictions = {}  # {anchor gene: cluster predictions}
     for i, anchor in enumerate(anchor_gene_names):
         logging.debug("Detecting cluster around anchor gene %r (%d of %d)", anchor, i + 1, len(anchor_gene_names))
-
-        try:
-            anchor_promoter_index = get_anchor_promoter_index(anchor, promoters)
-        except ValueError:
-            logging.debug("No promoter region for %r, skipping this anchor gene", anchor)
-            continue
-
-        # predict motifs with MEME ("de novo")
-        meme_dir = os.path.join(options.output_dir, "meme", anchor)
-        promoter_sets = get_promoter_sets(meme_dir, anchor_promoter_index, promoters)
-        exit_code = run_meme(meme_dir, options, VERBOSE_DEBUG)
-        if exit_code != 0:
-            logging.warning("MEME discovered a problem (exit code %d), skipping this anchor gene", exit_code)
-            continue
-        motifs = filter_meme_results(meme_dir, promoter_sets, anchor)
-
-        if not motifs:
-            logging.debug("Could not predict motifs around %r, skipping this anchor gene", anchor)
-            continue
-
-        # search motifs with FIMO ("scanning")
-        fimo_dir = os.path.join(options.output_dir, "fimo", anchor)
-        exit_code = run_fimo(meme_dir, fimo_dir, record, options, VERBOSE_DEBUG)
-        if exit_code != 0:
-            logging.warning("FIMO discovered a problem (exit code %d), skipping this anchor gene", exit_code)
-            continue
-        motifs = filter_fimo_results(motifs, fimo_dir, promoters, anchor_promoter_index)
-
-        if not motifs:
-            logging.debug("Could not find motif occurrences for %r, skipping this anchor gene", anchor)
-            continue
-
-        # TODO: as4, SiTaR (http://bioinformatics.oxfordjournals.org/content/27/20/2806):
-        # Alternative to MEME and FIMO. Part of the original CASSIS implementation.
-        # No motif prediction (no MEME). Motif search with SiTaR (instead if FIMO).
-        # Have to provide a file in FASTA format with binding site sequences of at least one transcription factor.
-        # Will result in binding sites per promoter (like FIMO) --> find islands
-        #
-        # implement: YES? NO?
-
-        # find islands of binding sites around anchor gene
-        islands = get_islands(anchor_promoter_index, motifs, promoters)
-        logging.debug("%d possible cluster predictions for %r", len(islands), anchor)
-
-        # return cluster predictions sorted by border abundance
+        # get cluster predictions sorted by border abundance
         # (most abundant --> "best" prediction)
-        cluster_predictions[anchor] = sort_by_abundance(islands)
-        cluster_predictions[anchor] = check_cluster_predictions(cluster_predictions[anchor],
-                                                                record, promoters, ignored_genes)
-
-        predicted_borders.extend(create_cluster_borders(anchor, cluster_predictions[anchor], record))
+        predictions = get_predictions_for_anchor(anchor, promoters, record, ignored_genes, options)
+        if predictions:
+            cluster_predictions[anchor] = predictions
+            predicted_borders.extend(create_cluster_borders(anchor, predictions, record))
 
     logging.debug("Cleaning up MEME and FIMO output directories")
     cleanup_outdir(anchor_gene_names, cluster_predictions, options)
@@ -355,447 +225,6 @@ def ignore_overlapping(genes):
         logging.debug("Ignoring %d genes due to overlapping locations", len(ignored))
 
     return (genes, ignored)
-
-
-def get_anchor_promoter_index(anchor: str, promoters: List[Promoter]) -> int:
-    """Find the name of the promoter which includes the anchor gene"""
-    # the promoter ID is not (necessarily) equal to the anchor ID!
-    for i, promoter in enumerate(promoters):
-        if anchor in promoter.get_gene_names():
-            return i
-
-    raise ValueError("No promoter exists for the given anchor: %s" % anchor)
-
-
-def get_promoter_sets(meme_dir: str, anchor_promoter: int, promoters: List[Promoter]) -> List[Motif]:
-    """Prepare sets of promoter sequences and motif subdirectories"""
-    promoter_sets = []
-
-    if not os.path.exists(meme_dir):
-        os.makedirs(meme_dir)
-
-    # prepare sets of promoter sequences (MEME input)
-    indices = set()  # type: Set[Tuple[int, int]] # to monitor unique start_index/end_index
-    for pm in PROMOTER_RANGE:
-        start_index = anchor_promoter - pm.minus
-        end_index = anchor_promoter + pm.plus
-
-        if start_index < 0:  # anchor promoter near beginning of record --> truncate
-            if VERBOSE_DEBUG:
-                logging.debug("Promoter set " + pm.pairing_string + " exceeds upstream record border")
-            start_index = 0
-
-        if end_index > len(promoters) - 1:  # anchor promoter near end of record --> truncate
-            if VERBOSE_DEBUG:
-                logging.debug("Promoter set " + pm.pairing_string + " exceeds downstream record border")
-            end_index = len(promoters) - 1
-
-        # discard promoter sets, which reappear due to truncation
-        if (start_index, end_index) not in indices:
-            indices.add((start_index, end_index))
-
-            # check (again, compare init of PROMOTER_RANGE) if the promoter set has at least 4 promoters
-            if end_index - start_index + 1 >= 4:
-                promoter_sets.append(Motif(pm.plus, pm.minus))
-
-                pm_dir = os.path.join(meme_dir, pm.pairing_string)
-                if not os.path.exists(pm_dir):
-                    os.makedirs(pm_dir)
-
-                # write promoter sequences to fasta file, in respective "plus-minus" subdir
-                with open(os.path.join(pm_dir, "promoters.fasta"), "w") as pm_handle:
-                    for i in range(start_index, end_index + 1):
-                        seq = SeqRecord(promoters[i].seq,
-                                        id=promoters[i].get_id(),
-                                        description="length={}bp".format(len(promoters[i].seq)))
-                        if i == anchor_promoter:  # mark anchor gene
-                            seq.id += "__ANCHOR"  # must be part of id, otherwise MEME woun't recognize it
-                        SeqIO.write(seq, pm_handle, "fasta")
-            else:
-                if VERBOSE_DEBUG:
-                    logging.debug("Too short promoter set " + pm.pairing_string)
-        else:
-            if VERBOSE_DEBUG:
-                logging.debug("Duplicate promoter set " + pm.pairing_string)
-
-    return promoter_sets
-
-
-def filter_meme_results(meme_dir: str, promoter_sets: List[Motif], anchor):
-    """Analyse and filter MEME results"""
-    for motif in promoter_sets:
-        xml_file = os.path.join(meme_dir, motif.pairing_string, "meme.xml")
-        e = ElementTree.parse(xml_file).getroot()
-        reason = e.find("model/reason_for_stopping").text
-        anchor_seq_id = ""
-
-        # no motif found for given e-value cutoff :-(
-        if "Stopped because motif E-value > " in reason:
-            if VERBOSE_DEBUG:
-                logging.debug("MEME: motif %s; e-value exceeds cutoff", motif.pairing_string)
-
-        # motif(s) found :-)
-        elif "Stopped because requested number of motifs (1) found" in reason:
-            # find anchor genes' sequence_id
-            training_set = e.findall("training_set/sequence")  # all promoter sequences passed to MEME
-            for element in training_set:
-                if "__ANCHOR" in element.attrib["name"]:
-                    anchor_seq_id = element.attrib["id"]  # e.g. id=sequence_1
-
-            # only accept motifs which occur in the anchor genes promoter
-            # sequences which contributed to the motif
-            contributing_sites = e.findall("motifs/motif/contributing_sites/contributing_site")
-            if anchor_seq_id in map(lambda site: site.attrib["sequence_id"], contributing_sites):
-                # save motif score
-                motif.score = float(e.find("motifs/motif").attrib["e_value"])  # one motif, didn't ask MEME for more
-
-                # save sequence sites which represent the motif
-                motif.seqs = ["".join(map(lambda letter: letter.attrib["letter_id"],
-                                          site.findall("site/letter_ref")))
-                              for site in contributing_sites]
-                # write sites to fasta file
-                with open(os.path.join(meme_dir, str(motif),
-                                       "binding_sites.fasta"), "w") as handle:
-                    handle.write(">{}__{}\n".format(anchor, str(motif)))
-                    handle.write("\n".join(motif.seqs))
-                if VERBOSE_DEBUG:
-                    logging.debug("MEME: motif %s; e-value = %s", motif, motif.score)
-            else:
-                if VERBOSE_DEBUG:
-                    logging.debug("MEME: motif %s; does not occur in anchor gene promoter", motif)
-
-        # unexpected reason, don't know why MEME stopped :-$
-        else:
-            logging.error("MEME stopped unexpectedly (reason: " + reason + ")")
-
-    return [motif for motif in promoter_sets if motif.score is not None]
-
-
-def filter_fimo_results(motifs: List[Motif], fimo_dir: str, promoters: List[Promoter],
-                        anchor_promoter: int) -> List[Motif]:
-    """Analyse and filter FIMO results"""
-
-    filtered = []
-
-    for motif in motifs:
-        assert isinstance(motif, Motif), type(motif)
-        with open(os.path.join(fimo_dir, str(motif), "fimo.txt"), "r") as handle:
-            table_reader = csv.reader(handle, delimiter="\t")
-            for row in table_reader:
-                # skip comment lines
-                if row[0].startswith("#"):
-                    continue
-                seq_id = row[1]
-                motif.hits[seq_id] += 1
-
-        # write binding sites per promoter to file
-        with open(os.path.join(fimo_dir, str(motif), "bs_per_promoter.csv"), "w") as handle:
-            table_writer = csv.writer(handle, delimiter="\t", lineterminator="\n")
-            table_writer.writerow(["#", "promoter", "binding sites"])  # table head
-            for i, promoter in enumerate(promoters):
-                table_writer.writerow([i+1, promoter.get_id(), motif.hits.get(promoter.get_id(), 0)])
-
-        percentage = len(motif.hits) / len(promoters) * 100
-        if percentage == 0.0:
-            # too low
-            if VERBOSE_DEBUG:
-                logging.debug("FIMO: motif %s; occurs in %d promoters (no hits)",
-                              motif, len(motif.hits))
-            continue
-        elif percentage > MAX_PERCENTAGE:
-            # too high
-            if VERBOSE_DEBUG:
-                logging.debug("FIMO: %s; occurs in %d promoters; %.2f%% of all promoters (too many)",
-                              motif, len(motif.hits), percentage)
-            continue
-        elif promoters[anchor_promoter].get_id() not in motif.hits:  # not in achor promoter
-            # no site in anchor promoter
-            if VERBOSE_DEBUG:
-                logging.debug("FIMO: motif %s; no hits in the promoter of the anchor gene", motif)
-            continue
-
-        # everything ok
-        if VERBOSE_DEBUG:
-            logging.debug("FIMO: motif %s; occurs in %d promoters; %.2f%% of all promoters",
-                          motif, len(motif.hits), percentage)
-        filtered.append(motif)
-
-    return filtered
-
-
-def get_island_gap_size(position: int, promoter_scores: List[int], max_gap_size: int,
-                        upstream: bool) -> int:
-    """ Finds the size of an island in promoters, if it exists.
-
-        Arguments:
-            position: the position around which to check for an island in promoter_scores
-            promoter_scores: the number of hits for each promoter, ordered by location
-            max_gap_size: the maximum gap size allowable for an island
-            upstream: whether the search is looking upstream or not
-
-        Returns:
-            0 if no island found, otherwise an int less than or equal to max_gap_size
-    """
-    if upstream:
-        end = position + 1
-    else:
-        start = position
-
-    for gap_size in range(1, max_gap_size + 1):
-        if upstream:
-            start = position - gap_size - 1
-        else:
-            end = position + gap_size + 2
-
-        # skip anything that there are no promoters for
-        if start < 0 or end > len(promoter_scores):
-            continue
-
-        # for gap_size of 1: x 0 x, where 0 is the anchor
-        # for gap_size of 2: x 0 0 x...
-        # and so on
-        chunk = promoter_scores[start:end]
-        if chunk[0] > 0 and chunk[-1] > 0 and chunk[1:-1] == [0] * gap_size:
-            # since no larger island can contain a smaller one, stop
-            return gap_size
-
-    return 0
-
-
-def get_islands(anchor_promoter: int, motifs: List[Motif], promoters: List[Promoter]):
-    """Find islands of binding sites (previously found by FIMO) around anchor gene to define cluster borders"""
-    assert MAX_GAP_LENGTH >= 0
-    islands = []
-    motifs = list(motifs)
-    for motif in motifs:
-        # create list with binding sites per promoter
-        bs_per_promoter = [0] * len(promoters)  # first: set number of binding sites to 0
-        for i, promoter in enumerate(promoters):
-            # second: set actual number of binding sites, if any
-            bs_per_promoter[i] = motif.hits[promoter.get_id()]
-
-        # upstream
-        start = anchor_promoter  # init upstream cluster border
-        i = anchor_promoter  # init position of anchor gene's promoter
-        while i > 0:
-            # promoter with binding site
-            # … 1 …
-            if bs_per_promoter[i-1] > 0:
-                start -= 1
-                i -= 1
-                continue
-            # check with gaps, e.g. 1 0 1 or 1 0 0 1
-            gap_size = get_island_gap_size(i, bs_per_promoter, MAX_GAP_LENGTH, upstream=True)
-            if gap_size == 0:
-                break
-            start -= gap_size + 1
-            i -= gap_size + 1
-
-        # downstream
-        i = anchor_promoter  # reset position of anchor gene's promoter
-        end = anchor_promoter  # init downstream cluster border
-        while i < len(bs_per_promoter) - 1:
-
-            # promoter with binding site(s)
-            # … 1 …
-            if bs_per_promoter[i + 1] > 0:
-                end += 1
-                i += 1
-                continue
-
-            # check with gaps, e.g. 1 0 1 or 1 0 0 1
-            gap_size = get_island_gap_size(i, bs_per_promoter, MAX_GAP_LENGTH, upstream=False)
-            if gap_size == 0:
-                break
-            end += gap_size + 1
-            i += gap_size + 1
-
-        if VERBOSE_DEBUG:
-            logging.debug("Island %s -- %s (motif %s)", promoters[start].get_id(),
-                          promoters[end].get_id(), motif)
-        islands.append(Island(promoters[start], promoters[end], motif))
-    return islands
-
-
-class ClusterMarker(Pairing):
-    """ Marks the start or end position for a cluster by gene and motif.
-        Tracks abundance and the motif string (+n/-m) for that position.
-    """
-    def __init__(self, gene: str, motif: Motif) -> None:
-        super().__init__(motif.plus, motif.minus)
-        self.gene = str(gene)
-        self.abundance = 1
-        self.score = float(motif.score)
-        self.promoter = None  # type: Optional[str]
-
-    def update(self, motif) -> None:
-        """ Uses the given motif's values instead of the old ones, if the given
-            motif has a better (lower) score.
-        """
-        if self.score > float(motif.score):
-            self.score = float(motif.score)
-            self.plus = int(motif.plus)
-            self.minus = int(motif.minus)
-
-    def __str__(self) -> str:
-        return "gene %s; abundance %s; motif %s; score %s" % (
-                  self.gene, self.abundance, self.pairing_string, self.score)
-
-    def __repr__(self) -> str:
-        return "ClusterMarker(%s, promoter=%s)" % (str(self), self.promoter)
-
-    def __eq__(self, other: Any) -> bool:
-        return (isinstance(other, ClusterMarker)
-                and all(getattr(self, key) == getattr(other, key) for key in vars(self)))
-
-
-class ClusterPrediction:
-    """ A prediction of a cluster. Contains start and end ClusterMarkers along with
-        counts of genes and promoters within that range.
-    """
-    def __init__(self, start: ClusterMarker, end: ClusterMarker) -> None:
-        self.start = start
-        self.end = end
-        self.genes = 0
-        self.promoters = 0
-
-    def __repr__(self) -> str:
-        return "ClusterPrediction(start=%s, end=%s, gene_count=%d, promoter_count=%d)" % (
-                        self.start, self.end, self.genes, self.promoters)
-
-    def __eq__(self, other: Any) -> bool:
-        return (isinstance(other, ClusterPrediction)
-                and all(getattr(self, key) == getattr(other, key) for key in vars(self)))
-
-
-def sort_by_abundance(islands: List[Island]) -> List[ClusterPrediction]:
-    """Sort upstream (start) and downstream (end) borders of islands by abundance"""
-    # count border abundance
-    # start/end are treated independently!
-    starts = {}  # type: Dict[str, ClusterMarker]
-    ends = {}  # type: Dict[str, ClusterMarker]
-    for island in islands:
-        if island.start.gene_name in starts:
-            scorer = starts[island.start.gene_name]
-            scorer.abundance += 1
-            scorer.update(island.motif)
-        else:
-            starts[island.start.gene_name] = ClusterMarker(island.start.gene_name, island.motif)
-
-        if island.end.get_gene_names()[-1] in ends:
-            scorer = ends[island.end.get_gene_names()[-1]]
-            scorer.abundance += 1
-            scorer.update(island.motif)
-        else:
-            ends[island.end.get_gene_names()[-1]] = ClusterMarker(island.end.get_gene_names()[-1], island.motif)
-
-    # compute sum of start and end abundance, remove duplicates, sort descending
-    abundances_sum_sorted = sorted(set([s.abundance + e.abundance
-                                        for s in starts.values() for e in ends.values()]), reverse=True)
-    # compute sum of start and end motif score, remove duplicates, sort ascending
-    scores_sum_sorted = sorted(set([s.score + e.score
-                                    for s in starts.values() for e in ends.values()]))
-    # sort by value (=abundance) of start, descending
-    starts_sorted = sorted(starts, key=lambda x: starts[x].abundance, reverse=True)
-    # sort by value (=abundance) of end, descending
-    ends_sorted = sorted(ends, key=lambda x: ends[x].abundance, reverse=True)
-
-    clusters = []
-    for abundance in abundances_sum_sorted:
-        # list from highest (best) to lowest (worst) abundance
-        for score in scores_sum_sorted:
-            # list from lowest (best) to highest (worst) motif score/e-value
-            for start in starts_sorted:
-                for end in ends_sorted:
-                    if (starts[start].abundance + ends[end].abundance == abundance
-                            and starts[start].score + ends[end].score == score):
-                        start_marker = starts[start]
-                        end_marker = ends[end]
-                        clusters.append(ClusterPrediction(start_marker, end_marker))
-                        if VERBOSE_DEBUG:
-                            logging.debug("Upstream border:   %s", start_marker)
-                            logging.debug("Downstream border: %s", end_marker)
-                            logging.debug("Total abundance %s, total score %.1e", abundance, score)
-    return clusters
-
-
-def check_cluster_predictions(cluster_predictions: List[ClusterPrediction],
-                              record: Record, promoters: List[Promoter], ignored_genes: List[Gene]):
-    """Get some more info about each cluster prediction and check if it is sane"""
-    checked_predictions = []
-    for cp, prediction in enumerate(cluster_predictions):
-        sane = True
-
-        # find indices of first and last GENE of the cluster prediction in all genes
-        all_gene_names = [gene.get_name() for gene in sorted(record.get_genes(), key=lambda x: x.location.start)]
-
-        start_index_genes = None
-        end_index_genes = None
-        for i, gene_name in enumerate(all_gene_names):
-            if not start_index_genes and prediction.start.gene == gene_name:
-                start_index_genes = i
-            if not end_index_genes and prediction.end.gene == gene_name:
-                end_index_genes = i
-            if start_index_genes and end_index_genes:
-                break
-
-        # find indices of first and last PROMOTER of the cluster prediction in all promoters
-        start_index_promoters = None
-        end_index_promoters = None
-        for i, promoter in enumerate(promoters):
-            if not start_index_promoters and prediction.start.gene in promoter.get_gene_names():
-                start_index_promoters = i
-            if not end_index_promoters and prediction.end.gene in promoter.get_gene_names():
-                end_index_promoters = i
-            if start_index_promoters and end_index_promoters:
-                break
-
-        prediction.start.promoter = promoters[start_index_promoters].get_id()
-        prediction.end.promoter = promoters[end_index_promoters].get_id()
-        prediction.genes = end_index_genes - start_index_genes + 1
-        prediction.promoters = end_index_promoters - start_index_promoters + 1
-        if VERBOSE_DEBUG:
-            if cp == 0:
-                logging.debug("Best prediction (most abundant): %r -- %r",
-                             prediction.start.gene, prediction.end.gene)
-            else:
-                logging.debug("Alternative prediction (%s): %r -- %r",
-                             cp, prediction.start.gene, prediction.end.gene)
-
-        # warn if cluster prediction right at or next to record (~ contig) border
-        if start_index_genes < 10:
-            if VERBOSE_DEBUG:
-                logging.debug("Upstream cluster border located at or next to sequence record border,"
-                              " prediction could have been truncated by record border")
-            sane = False
-        if end_index_genes > len(all_gene_names) - 10:
-            if VERBOSE_DEBUG:
-                logging.debug("Downstream cluster border located at or next to sequence record border,"
-                              " prediction could have been truncated by record border")
-            sane = False
-
-        # warn if cluster prediction too short (includes less than 3 genes)
-        if prediction.genes < 3:
-            if VERBOSE_DEBUG:
-                logging.debug("Cluster is very short (less than 3 genes). Prediction may be questionable.")
-            sane = False
-
-        # warn if ignored gene (overlapping with anthor gene, see ignore_overlapping())
-        # would have been part of the cluster
-        for ignored_gene in ignored_genes:
-            if ignored_gene.get_name() in all_gene_names[start_index_genes: end_index_genes + 1]:
-                if VERBOSE_DEBUG:
-                    logging.debug("Gene %r is part of the predicted cluster,"
-                                 " but it is overlapping with another gene and was ignored", ignored_gene)
-                    logging.debug("Gene %r could have affected the cluster prediction", ignored_gene)
-                # sane = False # uncomment if you want alternatives for predictions with ignored genes, too # TODO
-                break
-
-        checked_predictions.append(prediction)
-
-        if sane:  # TODO: this seems wrong
-            break
-    return checked_predictions
 
 
 def cleanup_outdir(anchor_gene_names: List[str], cluster_predictions: Dict[str, List[ClusterPrediction]], options):
